Add a `node_executable` to markdownlint toolchain
This makes the `markdownlint_aspect` work on my macOS system. It wasn't
previously working because in the Bazel action sandbox environment, the
`markdownlint` script wasn't finding the `node` executable.
Specifically, the sh-bang line

```
```

was failing because `node` wasn't found on the path.

Now, instead of executing the `markdownlint` executable, the Bazel
action executes the `node` executable which interprets the
`markdownlint` as a Node.js script.

This problem could have probably been solved (more simply) by instead
setting the action's `use_default_shell_env` to `True`, but I'd rather
keep as much of the user's shell environment away from the Bazel actions
as possible. I'd like to keep Bazel actions as hermetic as possible.