Draft stub for GitHub pages website
This stub website is built with `@dwtj_rules_hugo`.

This website was put in the `website/` directory as a nested repository.
It is a separate repository so that its the main repository doesn't
depend upon all of its external repositories. (E.g., users shouldn't
to import `@dwtj_rules_hugo` to use `@dwtj_rules_markdown`.)